rename files and set max client width on viewport
